move cloudfunctions into separate files
